refactor: huge refactor

+++ K3DGE/Render/src/main/kotlin/k3dge/render/renderer2d/model/SpriteSizeEnum.kt
@@ -1,9 +1,9 @@
 package k3dge.render.renderer2d.model
 
-enum class SpriteSizeEnum(val value: Int) {
-    X1(1),
-    X4(4),
-    X8(8),
-    X16(16),
-    X32(32)
+enum class SpriteSizeEnum(val value: Float) {
+    X1(1.0f),
+    X4(4.0f),
+    X8(8.0f),
+    X16(16.0f),
+    X32(32.0f)
 }
